Minor Updates
The active character's texture now ignores the light system, and the chaser's cone of vision is slightly widened. Also added a bogey square in the Load Maze scene so that the shadow compositor would stop being upset because there were no shadows to composite. The square is in the middle of the maze, but has a low enough Z value that it should not be at all visible.